+++ CHANGELOG.md
@@ -12,10 +12,15 @@ This project adheres to Semantic Versioning.
 
 * `window::get_physical_size` has been added, making it possible to find out the size of the window in unscaled, physical pixels.
 * `window::get_dpi_scale` has been added, allowing the window/monitor's pixel density to be queried. ([@sumibi-yakitori](https://github.com/sumibi-yakitori) in [#249](https://github.com/17cupsofcoffee/tetra/pull/249)
+* `Texture` and `Canvas` now have a `get_data` method, allowing their texture data to be copied from the GPU into an `ImageData`.
+* `ImageData` now has a `into_rgba8` method, which allows you to retrieve the underlying buffer.
 
 ### Changed
 
 * Canvases are now initialized to transparent black, rather than containing unspecified data on creation.
+* `ImageData::from_rgba` has been replaced with `ImageData::from_rgba8`, and `BmFontBuilder::with_page_rgba` has been replaced with `BmFontBuilder::with_page_rgba8`.
+    * This is for consistency with existing methods (e.g. on `Color`).
+    * The old names have been deprecated, and will be removed in 0.7.
 
 ### Fixed